Move capStoneSizeFields to test
- Remove capStoneAnisoAdapt which became capVol.
- Move capStoneSizeFields.h to test/capVolSizeFields.h.
- Update `#include` and CMake include_directories.
